requirements.txt
# RLdC Trading Bot - Python Dependencies
# Backend Framework
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.5
pydantic-settings==2.7.1

# Database
sqlalchemy==2.0.36
alembic==1.14.0

# Binance Integration
python-binance==1.0.21
ccxt==4.5.20
websockets==14.1

# Data Processing & Analysis
pandas==2.2.3
numpy==2.2.3
pandas-ta==0.3.14b0

# Telegram Bot
python-telegram-bot==21.10

# Utilities
python-dotenv==1.0.1
requests==2.32.3
aiohttp==3.11.11
python-dateutil==2.9.0.post0

# Logging & Monitoring
loguru==0.7.3

# Testing
pytest==8.3.4
pytest-asyncio==0.25.2
httpx==0.28.1

# CSV Export
openpyxl==3.1.5
